backend router division accomplished
